swap_face.py
import cv2
import numpy as np
from PIL import Image

def check_and_convert_orientation_conv_gray(image_path):
    """Convert image to grayscale and return as numpy array"""
    image = cv2.imread(image_path)
    if image is None:
        raise ValueError(f"Could not load image from {image_path}")
    gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
    return gray

def swap(image1, image2):
    """Basic face swap placeholder - returns the first image"""
    # This is a placeholder implementation
    # You'll need to implement actual face swapping logic here
    if isinstance(image1, np.ndarray):
        return Image.fromarray(image1)
    return image1