requirements.txt
dlib==19.24.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0
requests==2.31.0
streamlit==1.31.1
streamlit-cropper==0.2.1
